A program that solves this + is + very = easy
A challenge on a job application that asked for one to solve "this + is + very = easy" where each letter is asigned a unique number. I used a basic program that randomised numbers until it had a solution. Not a very elegant code but it was the first time I had written code in over 2 years, I enjoyed it and it spurred me to get back into coding.